Line Chart was set and designed, date type changed to long(timestamp), formatters for date and axises in chart made

+++ app/src/main/java/spryrocks/com/tristimer/presentation/ui/screens/statistics/StatisticsModel.java
@@ -3,15 +3,15 @@ package spryrocks.com.tristimer.presentation.ui.screens.statistics;
 import com.spryrocks.android.modules.ui.mvvm.model.ObservableField;
 
 public class StatisticsModel {
-    public final ObservableField<Float> best = new ObservableField<>();
-    public final ObservableField<Float> worst = new ObservableField<>();
-    public final ObservableField<Float> avg5 = new ObservableField<>();
-    public final ObservableField<Float> bestAvg5 = new ObservableField<>();
-    public final ObservableField<Float> avg12 = new ObservableField<>();
-    public final ObservableField<Float> bestAvg12 = new ObservableField<>();
-    public final ObservableField<Float> avg50 = new ObservableField<>();
-    public final ObservableField<Float> bestAvg50 = new ObservableField<>();
-    public final ObservableField<Float> avg100 = new ObservableField<>();
-    public final ObservableField<Float> bestAvg100 = new ObservableField<>();
-    public final ObservableField<Float> sessionMean = new ObservableField<>();
+    public final ObservableField<Long> best = new ObservableField<>();
+    public final ObservableField<Long> worst = new ObservableField<>();
+    public final ObservableField<Long> avg5 = new ObservableField<>();
+    public final ObservableField<Long> bestAvg5 = new ObservableField<>();
+    public final ObservableField<Long> avg12 = new ObservableField<>();
+    public final ObservableField<Long> bestAvg12 = new ObservableField<>();
+    public final ObservableField<Long> avg50 = new ObservableField<>();
+    public final ObservableField<Long> bestAvg50 = new ObservableField<>();
+    public final ObservableField<Long> avg100 = new ObservableField<>();
+    public final ObservableField<Long> bestAvg100 = new ObservableField<>();
+    public final ObservableField<Long> sessionMean = new ObservableField<>();
 }
